+++ qool_qml_project_setup.cmake
@@ -9,7 +9,8 @@ function(append_qml_dir _V_)
   list(REMOVE_DUPLICATES QML_DIRS)
   set(QML_IMPORT_PATH
       ${QML_DIRS}
-      CACHE STRING "qt qml folder" FORCE)
+      CACHE STRING "qt qml folder" FORCE
+  )
   message("Adding QML_IMPORT_PATH: ${_V_}")
 endfunction()
 
@@ -32,7 +33,8 @@ macro(load_qoolui_standard_options)
 
     set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
     set(QOOLUI_PLUGIN_OUTPUT_DIRECTORY
-        "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${QOOL_PLUGIN_DIR}")
+        "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${QOOL_PLUGIN_DIR}"
+    )
     set(QT_QML_OUTPUT_DIRECTORY ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/qml)
     append_qml_dir(${QT_QML_OUTPUT_DIRECTORY})
     set(QT_QML_GENERATE_QMLLS_INI ON)
@@ -63,41 +65,28 @@ macro(copy_qml_modules_for _T_)
     POST_BUILD
     COMMAND ${CMAKE_COMMAND} -E copy_directory ${QT_QML_OUTPUT_DIRECTORY}
             "$<TARGET_FILE_DIR:${_T_}>/qml"
-    COMMAND
-      ${CMAKE_COMMAND} -E copy_directory ${QOOLUI_PLUGIN_OUTPUT_DIRECTORY}
-      "$<TARGET_FILE_DIR:${_T_}>/qoolplugins")
+    COMMAND ${CMAKE_COMMAND} -E copy_directory
+            ${QOOLUI_PLUGIN_OUTPUT_DIRECTORY}
+            "$<TARGET_FILE_DIR:${_T_}>/qoolplugins"
+  )
 endmacro()
 
-function(convert_paths_to_relative VAR_NAME BASE_DIR)
-  # 获取原始变量的值（绝对路径列表）
-  set(ABSOLUTE_PATHS "${${VAR_NAME}}")
-  set(RELATIVE_PATHS "")
-
-  foreach(ABS_PATH ${ABSOLUTE_PATHS})
-    file(RELATIVE_PATH REL_PATH "${BASE_DIR}" "${ABS_PATH}")
-    list(APPEND RELATIVE_PATHS "${REL_PATH}")
-  endforeach()
-
-  # 直接修改调用者作用域中的变量
-  set(${VAR_NAME}
-      ${RELATIVE_PATHS}
-      PARENT_SCOPE)
-endfunction()
-
-function(qoolui_scan_assets_to OUTPUT_VAR)
+function(qoolui_collect_assets OUTPUT_VAR)
   set(ASS_DIR "${CMAKE_CURRENT_SOURCE_DIR}/assets")
   if(NOT EXISTS ${ASS_DIR})
     message(WARNING "assets directory not exists.")
-    set(${OUTPUT_VAR} "" PARENT_SCOPE})
+    set(${OUTPUT_VAR} "" PARENT_SCOPE)
     return()
   endif()
-  file(GLOB_RECURSE ASS_PATHS "${CMAKE_CURRENT_SOURCE_DIR}/assets/*")
-  convert_paths_to_relative(ASS_PATHS "${CMAKE_CURRENT_SOURCE_DIR}")
+  # RELATIVE 直接产出相对 CMAKE_CURRENT_SOURCE_DIR 的路径，免去二次转换
+  file(GLOB_RECURSE ASS_PATHS RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}"
+       "${CMAKE_CURRENT_SOURCE_DIR}/assets/*")
   message("qoolui assets found in ${CMAKE_CURRENT_SOURCE_DIR}:")
   foreach(_X_ ${ASS_PATHS})
     message("  ${_X_}")
   endforeach()
   set(${OUTPUT_VAR}
       ${ASS_PATHS}
-      PARENT_SCOPE)
+      PARENT_SCOPE
+  )
 endfunction()